Organize GameBoard components into subfolders
