+++ src/components/GameBoard/components/overlays/DebugMessage.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { GAME_CONSTANTS } from '../../../utils/Constants';
+import { GAME_CONSTANTS } from '../../../../utils/Constants';
 
 interface DebugMessageProps {
   message: string;
