+++ src/components/GameBoard/components/renderers/TowerDragVisualization.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
-import { useGameStore } from '../../../models/store';
-import { GAME_CONSTANTS } from '../../../utils/Constants';
-import type { TowerSlot } from '../../../models/gameTypes';
-import type { DragState, DropZoneState, DragFeedback } from '../types';
+import { useGameStore } from '../../../../models/store';
+import { GAME_CONSTANTS } from '../../../../utils/Constants';
+import type { TowerSlot } from '../../../../models/gameTypes';
+import type { DragState, DropZoneState, DragFeedback } from '../../types';
 
 interface TowerDragVisualizationProps {
   dragState: DragState;
