+++ src/components/GameBoard/components/screens/PreparationScreen.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { useGameStore } from '../../../models/store';
-import { GAME_CONSTANTS } from '../../../utils/Constants';
+import { useGameStore } from '../../../../models/store';
+import { GAME_CONSTANTS } from '../../../../utils/Constants';
 
 export const PreparationScreen: React.FC = () => {
   const {
