+++ src/components/GameBoard/components/tabs/StatisticsTab.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { useGameStore } from '../../../models/store';
-import { formatCurrency, formatProfessional, formatSmartPercentage } from '../../../utils/numberFormatting';
-import type { Tower } from '../../../models/gameTypes';
+import { useGameStore } from '../../../../models/store';
+import { formatCurrency, formatProfessional, formatSmartPercentage } from '../../../../utils/numberFormatting';
+import type { Tower } from '../../../../models/gameTypes';
 
 export const StatisticsTab: React.FC = () => {
   const { currentWave, enemiesKilled, totalEnemiesKilled, totalGoldSpent, towers, maxTowers, energy, maxEnergy, actionsRemaining, maxActions, energyEfficiency } = useGameStore();
